chore(sdk): update pulumiservice to 1.3.0 (#39)

+++ packages/sdks/pulumiservice/lib/src/index/team_stack_permission_args.dart
@@ -36,7 +36,7 @@ class TeamStackPermissionArgs {
   Map<String, dynamic> toMap() {
     return <String, dynamic>{
       'organization': organization,
-      'permission': pulumi.Input.mapInputValue<TeamStackPermissionScope, double>(permission, (value) => value.wireValue),
+      'permission': pulumi.Input.mapInputValue<TeamStackPermissionScope, int>(permission, (value) => value.wireValue),
       'project': project,
       'stack': stack,
       'team': team,
@@ -46,11 +46,10 @@ class TeamStackPermissionArgs {
   factory TeamStackPermissionArgs.fromMap(Map<String, dynamic> map) {
     return TeamStackPermissionArgs(
       organization: pulumi.Input.fromValue(map['organization'] as String),
-      permission: pulumi.Input.fromValue(TeamStackPermissionScope.fromValue(map['permission']! as double)),
+      permission: pulumi.Input.fromValue(TeamStackPermissionScope.fromValue(map['permission']! as int)),
       project: pulumi.Input.fromValue(map['project'] as String),
       stack: pulumi.Input.fromValue(map['stack'] as String),
       team: pulumi.Input.fromValue(map['team'] as String),
     );
   }
 }
-
